DEPgraphql/resource-service/requirements.txt
fastapi
uvicorn
SQLAlchemy
requests
strawberry-graphql